Enable fatal warnings and fix them all across codebase
- Delete `@deprecated` members
- Increase stack size for jvm to mitigate compiler stack overflow
- Introduce `IterableOnce` type alias to `TraversableOnce` in Scala 2.12
  to fix deprecation in Scala 2.13 while maintaining Scala 2.12
  compatibility
- Add `@nowarn("cat=deprecation")` in Scala 2 macro code

+++ ujson/src/ujson/JsVisitor.scala
@@ -30,11 +30,11 @@ trait JsVisitor[-T, +J] extends Visitor[T, J]{
   def visitInt32(i: Int, index: Int): J = visitFloat64(i, index)
   def visitInt64(i: Long, index: Int): J = {
     if (math.abs(i) > math.pow(2, 53) || i == -9223372036854775808L) visitString(i.toString, index)
-    else visitFloat64(i, index)
+    else visitFloat64(i.toDouble, index)
   }
   def visitUInt64(i: Long, index: Int): J = {
     if (i > math.pow(2, 53) || i < 0) visitString(java.lang.Long.toUnsignedString(i), index)
-    else visitFloat64(i, index)
+    else visitFloat64(i.toDouble, index)
   }
 
   def visitFloat64String(s: String, index: Int): J = {
